requirements.txt
torch
transformers
numpy
Pillow
tqdm
scikit-learn
networkx
pandas
pyyaml
matplotlib